feat: change headers structure, add team button, config.js and state.js
