+++ src/common/config.js
@@ -1,0 +1,3 @@
+export default {
+  SERVER_URL: 'https://afternoon-falls-25894.herokuapp.com',
+};
