+++ src/components/pages/authorization-page/tokenHandling/decodeToken.js
@@ -2,7 +2,7 @@ function extractTokenExpiration(token) {
   const base64Url = token.split('.')[1];
   const base64 = base64Url.replace(/-/g, '+').replace(/_/g, '/');
   const jsonPayload = decodeURIComponent(atob(base64).split('').map((c) => `%${(`00${c.charCodeAt(0).toString(16)}`).slice(-2)}`).join(''));
-  return JSON.parse(jsonPayload).exp;
+  return parseInt(JSON.parse(jsonPayload).exp, 10) * 1000;
 }
 
 export default extractTokenExpiration;
