--- src/components/pages/authorization-page/tokenHandling/formattedTime.js
@@ -1,6 +1,0 @@
-function timestamp(time) {
-  const date = new Date(time * 1000);
-  return date;
-}
-
-export default timestamp;
