+++ src/services/HttpService.js
@@ -1,16 +1,17 @@
+import config from '../common/config';
+import state from '../common/state';
+
 class HttpService {
   static async fetch(resource, init) {
-    const token = localStorage.getItem('token');
-
     Object.assign(init, {
       withCredentials: true,
       headers: {
-        Authorization: `Bearer ${token}`,
+        Authorization: `Bearer ${state.getToken()}`,
         Accept: 'application/json',
       },
     });
 
-    return window.fetch(`https://afternoon-falls-25894.herokuapp.com${resource}`, init);
+    return window.fetch(`${config.SERVER_URL}${resource}`, init);
   }
 }
