+++ src/services/userRequest.js
@@ -1,13 +1,12 @@
-const userRequest = async (user, path) => {
-  const rawResponse = await fetch(`https://afternoon-falls-25894.herokuapp.com/${path}`, {
-    method: 'POST',
-    headers: {
-      Accept: 'application/json',
-      'Content-Type': 'application/json',
-    },
-    body: JSON.stringify(user),
-  });
-  return rawResponse;
-};
+import config from '../common/config';
+
+const userRequest = async (user, path) => fetch(`${config.SERVER_URL}${path}`, {
+  method: 'POST',
+  headers: {
+    Accept: 'application/json',
+    'Content-Type': 'application/json',
+  },
+  body: JSON.stringify(user),
+});
 
 export default userRequest;
